API para controlar una clinica, maneja base de datos de empleados, doctores y pasientes, y sus respectivas autenticaciones
